Refactor file handling: move GridFS functions to utils and update import paths in tenderController
